pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nura"
version = "0.1.0"
description = "Nura - Universal Event-Driven AI Agent Platform"
requires-python = ">=3.12"
authors = [{name = "Mindstorm"}]
license = {text = "MIT"}
dependencies = [
    "pydantic~=2.10.6",
    "openai~=1.66.3",
    "tenacity~=9.0.0",
    "pyyaml~=6.0.2",
    "loguru~=0.7.3",
    "tiktoken~=0.9.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "click~=8.1.0",  # For CLI
    "pillow>=10.0.0",  # Image processing for vision capabilities
]

[project.optional-dependencies]
sandbox = ["docker~=7.1.0"]
bedrock = ["boto3~=1.37.18"]
websearch = [
    "googlesearch-python~=1.3.0",
    "duckduckgo_search~=7.5.3",
    "beautifulsoup4~=4.13.3",
    "requests~=2.32.3",
]
feishu = ["lark-oapi"]
all = [
    "docker~=7.1.0",
    "boto3~=1.37.18",
    "googlesearch-python~=1.3.0",
    "duckduckgo_search~=7.5.3",
    "beautifulsoup4~=4.13.3",
    "requests~=2.32.3",
    "lark-oapi",
]

[project.scripts]
nura = "nura.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nura*"]
